Move BlockEditor under UIKit directory
