+++ xcode/Subconscious/Shared/UIKit/BlockEditor/Cells/ErrorCell.swift
