+++ xcode/Subconscious/Shared/UIKit/BlockEditor/Cells/TextBlockCell.swift
@@ -1,5 +1,5 @@
 //
-//  QuoteBlockCell.swift
+//  TextBlockCell.swift
 //  BlockEditor
 //
 //  Created by Gordon Brander on 7/20/23.
@@ -8,21 +8,20 @@
 import UIKit
 
 extension BlockEditor {
-    class QuoteBlockCell:
+    class TextBlockCell:
         UICollectionViewCell,
         UITextViewDelegate,
-        Identifiable
+        UIComponentViewProtocol
     {
-        static let identifier = "QuoteBlockCell"
+        static let identifier = "TextBlockCell"
         
         var id: UUID = UUID()
         
         weak var delegate: TextBlockDelegate?
         
-        lazy var textView = UITextView(frame: .zero)
+        lazy var textView = SubtextTextView(frame: .zero)
         
         private lazy var divider = UIView.divider()
-        private lazy var quoteIndent = createQuoteIndent()
         
         private lazy var toolbar = UIToolbar.blockToolbar(
             upButtonPressed: { [weak self] in
@@ -84,7 +83,6 @@ extension BlockEditor {
             textView.inputAccessoryView = toolbar
             
             addSubview(textView)
-            addSubview(quoteIndent)
             addSubview(divider)
 
             NSLayoutConstraint.activate([
@@ -95,14 +93,6 @@ extension BlockEditor {
                 divider.leadingAnchor.constraint(equalTo: leadingAnchor),
                 divider.trailingAnchor.constraint(equalTo: trailingAnchor),
                 
-                quoteIndent.leadingAnchor.constraint(
-                    equalTo: leadingAnchor,
-                    constant: 4
-                ),
-                quoteIndent.widthAnchor.constraint(equalToConstant: 2),
-                quoteIndent.topAnchor.constraint(equalTo: topAnchor),
-                quoteIndent.bottomAnchor.constraint(equalTo: bottomAnchor),
-
                 bottomAnchor.constraint(equalTo: divider.bottomAnchor),
             ])
         }
@@ -111,12 +101,6 @@ extension BlockEditor {
             fatalError("init(coder:) has not been implemented")
         }
         
-        private func createQuoteIndent() -> UIView {
-            let view = UIView(frame: .zero)
-            view.translatesAutoresizingMaskIntoConstraints = false
-            return view
-        }
-
         func render(_ state: TextBlockModel) {
             self.id = state.id
             if textView.text != state.text {
