+++ xcode/Subconscious/Shared/UIKit/BlockEditor/Models/TextBlockModel.swift
